weezly_shop/requirements.txt
asgiref==3.4.1
Django==4.0
django-crispy-forms==1.13.0
Pillow==9.0.0
sqlparse==0.4.2